Made the following changes: - Uploaded the sample result file - Takes the paths input from user - Analysation of the data also works - Creates a new sheet for each subject - Separated the functions into a new file
